Move isErp to erp.js.
