+++ src/util.js
@@ -165,14 +165,6 @@ function getOpt(optObject, option, defaultValue) {
       defaultValue;
 }
 
-function isErp(x) {
-  return x && _.isFunction(x.score) && _.isFunction(x.sample);
-}
-
-function isErpWithSupport(x) {
-  return isErp(x) && _.isFunction(x.support);
-}
-
 module.exports = {
   copyObj: copyObj,
   cpsForEach: cpsForEach,
@@ -192,7 +184,5 @@ module.exports = {
   std: std,
   getOpt: getOpt,
   sum: sum,
-  asArray: asArray,
-  isErp: isErp,
-  isErpWithSupport: isErpWithSupport
+  asArray: asArray
 };
