+++ tests/test-examples.js
@@ -3,6 +3,7 @@
 var _ = require('underscore');
 var fs = require('fs');
 var webppl = require('../src/main.js');
+var erp = require('../src/erp');
 
 var examplesDir = './examples/';
 
@@ -28,8 +29,8 @@ var generateTestCases = function() {
   _.each(examples, function(example) {
     exports[example] = function(test) {
       test.doesNotThrow(function() {
-        webppl.run(loadExample(example), function(s, erp) {
-          test.ok(util.isErp(erp));
+        webppl.run(loadExample(example), function(s, val) {
+          test.ok(erp.isErp(val));
         });
       });
       test.done();
